requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
freezegun==1.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mock==4.0.3
more-itertools==8.7.0
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-freezegun==0.4.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
pyflakes==2.2.0
flask-cors==3.0.10
boto3==1.17.27
botocore~=1.20.27
moto==2.0.2
authlib==0.15.3
requests==2.25.1
python-jose==3.2.0